Refactor imports and move JSON text extraction logic to ChatUtils for better organization
